feat: integração - report

+++ src/main/java/com/carlos/charles_api/dto/response/ServiceOrderStatistcsDTO.java
@@ -1,11 +1,11 @@
 package com.carlos.charles_api.dto.response;
 
 public record ServiceOrderStatistcsDTO(
-        Integer open,
-        Integer assigned,
-        Integer canceled,
-        Integer completed,
-        Integer closed,
-        Integer all
+        long open,
+        long assigned,
+        long canceled,
+        long completed,
+        long closed,
+        long all
 ) {
 }
